moved distributor edit form to dedicated component
