+++ apps/waterfall/waterfall/src/app/dashboard/statement/edit/edit.component.scss
@@ -3,31 +3,3 @@ footer {
   justify-content: center;
   margin: 32px 0;
 }
-
-mat-tab-group {
-  margin-top: 24px;
-}
-
-section.surface {
-  margin: 48px 24px 0 24px;
-
-  h3 {
-    font-weight: 800;
-    font-size: 26px;
-    line-height: 24px;
-    color: var(--foreground-third-text);
-  }
-
-  form.group {
-    display: flex;
-    flex-wrap: wrap;
-    column-gap: 32px;
-    * {
-      flex: 1;
-    }
-  }
-
-  .group-label {
-    cursor: pointer;
-  }
-}
