+++ libs/waterfall/src/lib/components/statement-distributor/edit/edit.component.scss
@@ -1,0 +1,27 @@
+mat-tab-group {
+  margin-top: 24px;
+}
+
+section.surface {
+  margin: 48px 24px 0 24px;
+
+  h3 {
+    font-weight: 800;
+    font-size: 26px;
+    line-height: 24px;
+    color: var(--foreground-third-text);
+  }
+
+  form.group {
+    display: flex;
+    flex-wrap: wrap;
+    column-gap: 32px;
+    * {
+      flex: 1;
+    }
+  }
+
+  .group-label {
+    cursor: pointer;
+  }
+}
